requirements.txt
# ComfyForge dependencies
pyyaml>=6.0
httpx>=0.27
rich>=13.0
typer>=0.12
websockets>=12.0
firebase-admin>=6.0
huggingface-hub>=0.23
aiofiles>=24.0
uvicorn>=0.30
fastapi>=0.111
jinja2>=3.1
pydantic>=2.0
tqdm>=4.66
Pillow>=10.0
